pyproject.toml
[project]
name = "vibe-git"
version = "0.1.0"
description = "Friction-free git workflows for vibe coding - MCP server"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastmcp>=2.8.0",
    "watchdog>=6.0.0",
    "beartype>=0.19.0",
    "plum-dispatch>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
    "cosmic-ray>=8.3.0",
]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]



[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "B008",  # do not perform function calls in argument defaults
    "B017",  # pytest.raises(Exception) is too generic
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "cosmic-ray>=8.3.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]

[project.scripts]
vibe-git-mcp = "vibe_git.main:main"
